_archive/projects/fullpotential_ai/fullpotential_core/agents/services/i-match-automation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.18.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.4
jinja2==3.1.3
